Fixversions extracted

+++ python/spacewalk/server/vex/vex/downloader/downloader.py
@@ -47,21 +47,22 @@ class Downloader():
         response = requests.get(url)
         if response.status_code == 200:
             return response.text.splitlines()
-        else:
-            raise Exception(f"No se pudo descargar el índice. Código de estado: {response.status_code}")
+
+        raise Exception(f"No se pudo descargar el índice. Código de estado: {response.status_code}")
 
     def parse_index_lines(self, lines):
-        """Parsea las líneas del índice para obtener nombres de archivos y sus hashes"""
+        """
+        Parsea las líneas del índice para obtener nombres de archivos y sus hashes
+        """
+        
         files = []
         for line in lines:
-            print(line)
-            if line.strip():  # Ignorar líneas vacías
-                parts = line.split()
-                if len(parts) >= 2:  # Al menos hash y nombre de archivo
-                    file_hash = parts[0]
-                    filename = parts[-1]  # El último elemento es el nombre del archivo
-                    files.append((filename, file_hash))
+            
+            parts = line.split()
+            if len(parts) >= 2:
+                files.append((parts[-1], parts[0]))  # (filename, hash)
         return files
+
 
     def retrieve_hash(self, filename):
         """
@@ -71,52 +72,79 @@ class Downloader():
         response = requests.get(url, stream=True)
         return response.content.decode().split()[0]
 
-    def check_hash_db(self, filename):
+    def check_hash_db(self, cve_id):
 
-        cve = filename.replace(".json", "")
+        try:
+            logging.critical(f"Checking hash for {cve_id}")
+            self.db_manager.connect()
+            result, status = self.db_manager.get_cve_hash(cve_id)
+        
+        except Exception as e:
+            logging.critical(f"{str(e)}")
+            
+        finally:
+            self.db_manager.close()
+
+        logging.critical(f"DB hash {result} , {status}")
+
+        return result, status
+    
+    def insert_hash_db(self, cve_id, file_hash, hash_status, hash_type='SHA256'):
+
+        
 
         try:
             self.db_manager.connect()
-            result = self.db_manager.get_cve_hash(cve, 'SHA256')
+            logging.critical(self.db_manager.insert_cve(cve_id))
+            self.db_manager.insert_vex_hash(cve_id, file_hash, hash_status, hash_type)
+            
+        except Exception as e:
+            logging.error(f"{str(e)}")
+
         finally:
             self.db_manager.close()
-        
-        return result
 
+
+    # TODO: Change returns by raise
     def download_file(self, filename):
         """
         Download a file.
         """
 
-        # Debug info
-        logging.debug(f"\nFile: {filename}")
         file_hash = self.retrieve_hash(filename)
-        logging.debug(f"Hash SHA-256: {file_hash}")
 
-        db_hash = self.check_hash_db(filename)
+        cve_id = filename.rstrip(".json")
+        db_hash, db_hash_status = self.check_hash_db(cve_id)
 
-        if db_hash == -1 or db_hash != file_hash:
+        logging.info(f"File hash: {file_hash}")
+        logging.info(f"DB hash: {db_hash}")
+        logging.info(f"DB hash status: {db_hash_status}")
 
-            logging.debug(f"File needs to be updated")
+        if str(db_hash) != str(file_hash):
+            logging.info(f"File needs to be updated")
 
             url = urljoin(self.BASE_URL, filename)
             local_path = os.path.join(self.DOWNLOAD_DIR, filename)
             
-            logging.debug(f"Descargando desde {url}...")
+            logging.info(f"Downloading from {url}...")
             response = requests.get(url, stream=True)
             
             if response.status_code == 200:
                 with open(local_path, 'wb') as f:
                     for chunk in response.iter_content(chunk_size=8192):
                         f.write(chunk)
-                logging.debug(f"Descarga completada: {local_path}")
+                logging.info(f"Downloading completed: {local_path}")
+
                 return file_hash
+            
             else:
                 logging.error(f"Error al descargar {filename}. Código de estado: {response.status_code}")
                 return None
-        
-        else:
-            logging.debug(f"File already exists, nothing to do here :)")
+
+        elif db_hash_status != 0:
+            logging.warning(f"File not valid, nothing to do here :)")
+            return None
+            
 
     def download(self):
         try:
@@ -131,35 +159,48 @@ class Downloader():
             downloaded_count = 0
             total_files = len(index_lines)
 
-            logging.debug(f"\nEncontrados {total_files} archivos en el índice.")
+            logging.debug(f"\nFound {total_files} files in index.")
             
             for filename in index_lines:
                 hash = self.download_file(filename)
-                if hash != None:
 
+                # Formatting CVE 
+                cve_id = filename.rstrip(".json")
+                
+
+                if hash != None:
                     try:
                         downloaded_count += 1
 
                         self.parser.parse(self.DOWNLOAD_DIR+filename)
                         logging.info(f"Parsing info with hash {hash}")
-                        self.parser.persist_data(hash, 'SHA256')
-                        self.remove_file(self.DOWNLOAD_DIR+filename)
+                        self.parser.persist_data()
 
-                        # Cálculos de progreso
-                        elapsed_time = time.time() - start_time
-                        progress = (downloaded_count / total_files) * 100
-                        if downloaded_count > 0:
-                            remaining_time = (elapsed_time / downloaded_count) * (total_files - downloaded_count)/60
-                        else:
-                            remaining_time = 0
-                        
-                        logging.info(f"\n▶ Progress: {downloaded_count}/{total_files} ({progress:.2f}%)")
-                        logging.info(f"▶ Time spent: {elapsed_time:.2f}s")
-                        logging.info(f"▶ Remaining: {remaining_time:.2f}min")
+                        logging.info(f"Inserting hash: {hash} for {cve_id}")
+                        self.insert_hash_db(cve_id, hash, 0, 'SHA256')
+
+                        self.remove_file(self.DOWNLOAD_DIR+filename)
                     
                     except Exception as e:
                         logging.error(f"Error while parsing {filename}: {str(e)}")
+                        if "product_tree" in str(e):
+                            self.insert_hash_db(cve_id, hash, -1, 'SHA256')
+                            self.remove_file(self.DOWNLOAD_DIR+filename)
                 
+                # Cálculos de progreso
+                elapsed_time = time.time() - start_time
+                progress = (downloaded_count / total_files) * 100
+                if downloaded_count > 0:
+                    remaining_time = (elapsed_time / downloaded_count) * (total_files - downloaded_count)/60
+                else:
+                    remaining_time = 0
+                
+                logging.info(f"""\n
+                                ▶ Progress: {downloaded_count}/{total_files} ({progress:.2f}%)
+                                ▶ Time spent: {elapsed_time:.2f}s
+                                ▶ Remaining: {remaining_time:.2f}min
+                            """)
+
             total_time = time.time() - start_time
             logging.info(f"\n[*] Process completed in {total_time:.2f} seg.")
             logging.info(f"[*] Downloaded files: {downloaded_count}/{total_files}")
